♻️ refactor(sync): migrate collections to descriptor catalog
+ document shelf_books tombstone-revision divergence
