+++ sharedLogic/src/jvmTest/kotlin/com/calypsan/listenup/client/data/sync/CollectionSyncHandlerTest.kt
@@ -3,13 +3,11 @@ package com.calypsan.listenup.client.data.sync
 import com.calypsan.listenup.api.dto.SharePermission
 import com.calypsan.listenup.api.sync.CollectionBookSyncPayload
 import com.calypsan.listenup.api.sync.CollectionShareSyncPayload
-import com.calypsan.listenup.api.sync.CollectionSyncPayload
 import com.calypsan.listenup.api.sync.SyncEvent
 import com.calypsan.listenup.client.data.local.db.ListenUpDatabase
 import com.calypsan.listenup.client.data.local.db.RoomTransactionRunner
 import com.calypsan.listenup.client.data.sync.handlers.CollectionBookSyncDomainHandler
 import com.calypsan.listenup.client.data.sync.handlers.CollectionShareSyncDomainHandler
-import com.calypsan.listenup.client.data.sync.handlers.CollectionSyncDomainHandler
 import com.calypsan.listenup.client.test.db.createInMemoryTestDatabase
 import com.calypsan.listenup.api.result.AppResult
 import io.kotest.core.spec.style.FunSpec
@@ -19,92 +17,15 @@ import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.coroutines.test.runTest
 
 /**
- * Covers the three Collections sync handlers (Collections — Room v24):
- * [CollectionSyncDomainHandler] (leaf), [CollectionBookSyncDomainHandler] (junction),
- * and [CollectionShareSyncDomainHandler] (leaf). Each exercises Room write-through for
- * Created/Updated upserts, Deleted tombstones, and catch-up tombstone application.
+ * Covers the two remaining Collections-family sync handlers (Collections — Room v24):
+ * [CollectionBookSyncDomainHandler] (junction) and [CollectionShareSyncDomainHandler]
+ * (leaf). Each exercises Room write-through for Created/Updated upserts, Deleted
+ * tombstones, and catch-up tombstone application. (The `collections` leaf domain moved
+ * to [com.calypsan.listenup.client.data.sync.domains.collectionsDomain] — see
+ * `CollectionsDomainTest`.)
  */
 class CollectionSyncHandlerTest :
     FunSpec({
-
-        // ── collections (leaf) ──────────────────────────────────────────────────
-
-        test("collection Created event upserts the row into Room") {
-            withCollectionHandler { handler, db ->
-                handler
-                    .onEvent(createdCollection(collectionPayload("c1", name = "Favourites")), isOwnEcho = false)
-                    .shouldBeInstanceOf<AppResult.Success<Unit>>()
-                val row = db.collectionDao().getById("c1")
-                row shouldNotBe null
-                row!!.name shouldBe "Favourites"
-                row.libraryId shouldBe "lib1"
-                row.ownerId shouldBe "u1"
-                row.revision shouldBe 1L
-                row.deletedAt shouldBe null
-            }
-        }
-
-        test("collection Updated event overwrites the existing row") {
-            withCollectionHandler { handler, db ->
-                handler.onEvent(createdCollection(collectionPayload("c1", name = "Favourites")), isOwnEcho = false)
-                handler.onEvent(
-                    SyncEvent.Updated(
-                        id = "c1",
-                        revision = 2L,
-                        occurredAt = 200L,
-                        clientOpId = null,
-                        payload = collectionPayload("c1", name = "Top Picks", revision = 2L),
-                    ),
-                    isOwnEcho = false,
-                )
-                val row = db.collectionDao().getById("c1")!!
-                row.name shouldBe "Top Picks"
-                row.revision shouldBe 2L
-            }
-        }
-
-        test("collection Deleted event soft-deletes the row") {
-            withCollectionHandler { handler, db ->
-                handler.onEvent(createdCollection(collectionPayload("c1")), isOwnEcho = false)
-                handler
-                    .onEvent(SyncEvent.Deleted(id = "c1", revision = 2L, occurredAt = 500L), isOwnEcho = false)
-                    .shouldBeInstanceOf<AppResult.Success<Unit>>()
-                db.collectionDao().getById("c1") shouldBe null
-            }
-        }
-
-        test("collection onCatchUpItem live item upserts the row") {
-            withCollectionHandler { handler, db ->
-                handler
-                    .onCatchUpItem(collectionPayload("c2", name = "Sci-Fi"), isTombstone = false)
-                    .shouldBeInstanceOf<AppResult.Success<Unit>>()
-                db.collectionDao().getById("c2")!!.name shouldBe "Sci-Fi"
-            }
-        }
-
-        test("collection onCatchUpItem tombstone soft-deletes the row") {
-            withCollectionHandler { handler, db ->
-                handler.onCatchUpItem(collectionPayload("c1"), isTombstone = false)
-                handler
-                    .onCatchUpItem(
-                        collectionPayload("c1", revision = 3L, deletedAt = 700L),
-                        isTombstone = true,
-                    ).shouldBeInstanceOf<AppResult.Success<Unit>>()
-                db.collectionDao().getById("c1") shouldBe null
-            }
-        }
-
-        test("collection handler self-registers under domainName 'collections'") {
-            val registry = ClientSyncDomainRegistry()
-            val db = createInMemoryTestDatabase()
-            try {
-                val handler = CollectionSyncDomainHandler(db, RoomTransactionRunner(db), registry)
-                handler.domainName shouldBe "collections"
-                registry.lookup("collections") shouldBe handler
-            } finally {
-                db.close()
-            }
-        }
 
         // ── collection_books (junction) ─────────────────────────────────────────
 
@@ -243,16 +164,6 @@ class CollectionSyncHandlerTest :
 
 // ── Helpers ─────────────────────────────────────────────────────────────────────
 
-private fun withCollectionHandler(block: suspend (CollectionSyncDomainHandler, ListenUpDatabase) -> Unit) =
-    runTest {
-        val db = createInMemoryTestDatabase()
-        try {
-            block(CollectionSyncDomainHandler(db, RoomTransactionRunner(db), ClientSyncDomainRegistry()), db)
-        } finally {
-            db.close()
-        }
-    }
-
 private fun withCollectionBookHandler(block: suspend (CollectionBookSyncDomainHandler, ListenUpDatabase) -> Unit) =
     runTest {
         val db = createInMemoryTestDatabase()
@@ -272,31 +183,6 @@ private fun withCollectionShareHandler(block: suspend (CollectionShareSyncDomain
             db.close()
         }
     }
-
-private fun collectionPayload(
-    id: String,
-    name: String = "My Collection",
-    revision: Long = 1L,
-    deletedAt: Long? = null,
-) = CollectionSyncPayload(
-    id = id,
-    libraryId = "lib1",
-    ownerId = "u1",
-    name = name,
-    isInbox = false,
-    revision = revision,
-    updatedAt = 100L,
-    deletedAt = deletedAt,
-)
-
-private fun createdCollection(payload: CollectionSyncPayload) =
-    SyncEvent.Created(
-        id = payload.id,
-        revision = payload.revision,
-        occurredAt = payload.updatedAt,
-        clientOpId = null,
-        payload = payload,
-    )
 
 private fun junctionPayload(
     collectionId: String,
